move send file button to top
